fix cyclical deps by making types files
