+++ backend-api/src/functions/asyncBiometricToken/handlerDependencies.ts
@@ -4,7 +4,7 @@ import {
   getBiometricToken,
   GetBiometricToken,
 } from "./getBiometricToken/getBiometricToken";
-import { SessionRegistry } from "../common/session/SessionRegistry";
+import { SessionRegistry } from "../common/session/SessionRegistry/SessionRegistry";
 import { DynamoDbAdapter } from "../adapters/aws/dynamo/dynamoDbAdapter";
 import { EventService } from "../services/events/eventService";
 import { IEventService } from "../services/events/types";
