+++ backend-api/src/functions/asyncFinishBiometricSession/handlerDependencies.ts
@@ -1,4 +1,4 @@
-import { SessionRegistry } from "../common/session/SessionRegistry";
+import { SessionRegistry } from "../common/session/SessionRegistry/SessionRegistry";
 import { DynamoDbAdapter } from "../adapters/aws/dynamo/dynamoDbAdapter";
 import { EventService } from "../services/events/eventService";
 import { IEventService } from "../services/events/types";
