+++ backend-api/src/functions/asyncTxmaEvent/handlerDependencies.ts
@@ -1,5 +1,5 @@
 import { DynamoDbAdapter } from "../adapters/aws/dynamo/dynamoDbAdapter";
-import { SessionRegistry } from "../common/session/SessionRegistry";
+import { SessionRegistry } from "../common/session/SessionRegistry/SessionRegistry";
 import { EventService } from "../services/events/eventService";
 import { IEventService } from "../services/events/types";
